Initial stubs for client implementation code.
Include a substantial portion of the protocol documentation.  Add stubs for
client implementations that map to all message types.  They may not all be
fully implemented (or implemented at all), but the documentation helps to
drive the implementation.

Change the Makefile format for SRCS to avoid having to edit previous lines.
Sort the files by name.