Implement Rerror client support.
Use it in parsing the response payload for Rversion.  It takes a pointer to
the message so it can consume it if the message is either Rerror or
otherwise not the expected message type.  In the former case, the message
encded errno is returned.  In the latter, EINVAL is returned.