Implement Tattach; fix umount.
Tattach for 9P2000.u (py9p implementation) requires uid[4] for Tattach, so
implement that even though it is not part of the public (expired) 9P2000.u
specification.  Default to root/0 for now.

To fix mounting, implement p9fs_root() as a stub that simply hashes the root
path and uses vfs_hash_get/vfs_hash_insert for the root vnode.

To fix unmounting, temporarily reset the return codes from EINVAL to 0 for
the inactive and reclaim vnops.