WIP implementing readdir()/open()/lookup() etc.
This required a relatively large changeset to implement, which would be
difficult to break up.  It's not done yet, as there are other issues that
this revealed which need to be addressed.

Implement the following vnode ops:
- VOP_READDIR: This requires a Tread on the directory, and returns a Rread
  message composed of stat objects.  This is similar to Tstat, so in turn
  this requires making Tstat's subroutines more generally available.  It
  also requires changing p9fs_client_{read,write}() to take I/O callbacks,
  so that the caller can use those subroutines, since they require operating
  on the original message, which are generated using mbufs, not uios.
- VOP_LOOKUP: This uses Twalk in combination with p9fs_nget() to return the
  vnode for the child object.
- VOP_OPEN: This uses Topen with the appropriate fid and caller mode.  If
  this succeeds, the vnode's VM object will be created.  If the vnode being
  used is a directory, this will clone the current fid in order to ensure
  that the primary fid is still usable for cloning.  (py9p doesn't allow
  cloning open fids, which isn't a standard-specified behavior.)
- VOP_CLOSE: Currently clunks the above-mentioned 'open fid' if open.
- VOP_ACCESS: Rudimentary Unix permissions implementation based on the
  vnode's vattr, generated via p9fs_client_stat().
- VOP_GETATTR: Implemented entirely via p9fs_client_stat().

Implement the following VFS op:
- VFS_STATFS: 9P2000{,.u} do not have a call that supports fetching this
  metadata, so fake it for now.  We'll need to support 9P2000.L for this.
  This allows the current implementation to support e.g. 'df' commands.

Various other updates:
- Change VOP_RECLAIM to ensure the vnode's VM object is destroyed, and if
  necessary to release any valid fids.
- Fix several issues in p9fs_nget():
  - Change to take a p9fs_session so that the mount object can be provided
    by vnode ops, which won't otherwise have access to it.
  - Change to call p9fs_client_stat() if it is constructing a new vnode,
    so that the vnode type is correct.
  - Actually provide the p9fs_node object via v_data so vnops can access it.
- Fix p9fs_mount() to set up the root vnode more correctly.
- Add several more P9 data types to p9fs_proto.h.
- Change all other vnode ops to print a message, so we can tell if they are
  being called and (probably) need to be implemented.
- Change p9fs_client_walk() to allow cloning fids by specifying a NULL
  namestr.  Also change it to require a qid pointer to copy the result to.
- Allow sources to be mounted at /usr/src.  This allows building without
  having to specify SYSDIR/WORLDDIR.

Add a large comment explaining the next steps in p9fs_open().  The gist is
that this needs to be more complete in terms of how fids are generated and
handled.  For one thing, Tattach has to be emitted for every user so that
the server knows who is making the requests based on the fid.  For another,
this means we need to Tclunk fids in p9fs_close(), not p9fs_reclaim() or
even p9fs_inactive(), because the latter two are global to all users.
Otherwise, in heavy use a given filesystem may run out of fids too easily.

Another TODO is to build in tolerance for protocol errors by changing
p9fs_msg_* and everything that builds on it to percolate errors up, most
likely eventually returning EPROTO.

Finally, a lot of this code needs to be cleaned up.  There's considerable
duplication in the client protocol implementations, to the point that it's
probably worth changing p9fs_msg_send() to take a callback that handles
the success case, then changing all the APIs to supply such a callback.